Convert date to ordinal calendar date

toOrdinalDate(
        date=NULL,
        language="English")

Arguments

date
Either NULL (the default which uses the current date returned via Sys.date() or a character string of the form YYYY-MM-DD.

language
OPTIONAL. Language (defaults to English) indicating what language rules to use in conversion to ordinal number.

Details

Typical use of the function is to call the function with any argument to get the current date in long, ordinal form. See examples.

Value

Function returns the character string with the long, ordinal calendar date.

Examples

toOrdinalDate()

#> list()

toOrdinalDate("2017-10-9") ## October 9th, 2017

#> [1] "October 9th, 2017"

toOrdinalDate("2017-10-09") ## October 9th, 2017

#> [1] "October 9th, 2017"

toOrdinalDate(c("2022-11-21", "2019-1-3", "1976-2-2", "1965-2-18"))

#> [1] "November 21st, 2022" "January 3rd, 2019"   "February 2nd, 1976" 
#> [4] "February 18th, 1965"

## "November 21st, 2022" "January 3rd, 2019" "February 2nd, 1976" "February 18th, 1965"